Added frontend UI

+++ backend/controllers/goalController.js
@@ -1,13 +1,14 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 const { text } = require('express')
 
 //@desc get Goals
 //@route GET /api/goals
 //access Private
 const getGoals = asyncHandler(async (req,res) => {
-  const goals = await Goal.find()
+  const goals = await Goal.find({user: req.user.id })
 
     res.status(200).json(goals)
 })
@@ -22,7 +23,8 @@ const setGoals = asyncHandler(async (req,res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     })
 
     res.status(200).json(goal)
@@ -38,6 +40,21 @@ const updateGoal = asyncHandler(async (req,res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    //Cheack for user
+    if(!user){
+        res.status(401)
+        throw new Error('User Not Found')
+    }
+
+    //Make Sure the loggedin user matches the goal user
+    if(goal.user.toString() !== user.id)
+    {
+        res.status(401)
+        throw new Error('user not Authorized')
     }
 
     const updateGoal=await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -58,6 +75,24 @@ const deleteGoals = asyncHandler(async (req,res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    //Cheack for user
+    if(!user){
+        res.status(401)
+        throw new Error('User Not Found')
+    }
+
+    //Make Sure the loggedin user matches the goal user
+    if(goal.user.toString() !== user.id)
+    {
+        res.status(401)
+        throw new Error('user not Authorized')
+    }
+
+
+    
     await goal.deleteOne()
     res.status(200).json({ id: req.params.id})
 })
